pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "specred"
version = "0.1.0"
description = "자연어 요구사항을 BDD 시나리오로, 시나리오를 실행 가능한 테스트 코드로. TDD의 Red 단계를 자동화합니다."
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "anthropic>=0.25",
    "google-genai>=1.0",
    "openai>=1.0",
    "watchdog>=4.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
specred = "specred.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["specred*"]
